% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_suezalla.R
\name{scale_colour_suezalla}
\alias{scale_colour_suezalla}
\title{ggplot2 colour scale using suezalla_palette}
\usage{
scale_colour_suezalla(
  palette_name = "polaroid",
  n = NULL,
  type = "discrete",
  ...
)
}
\arguments{
\item{palette_name}{Name of the palette (default = "polaroid")}

\item{n}{Number of colours}

\item{type}{"discrete" or "continuous"}

\item{...}{Additional arguments to ggplot2 scale functions}
}
\description{
ggplot2 colour scale using suezalla_palette
}
