% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_suezalla.R, R/theme_suezalla.R
\name{scale_fill_suezalla}
\alias{scale_fill_suezalla}
\title{Custom ggplot2 fill scale for suezalla palettes}
\usage{
scale_fill_suezalla(...)

scale_fill_suezalla(...)
}
\description{
Custom ggplot2 fill scale for suezalla palettes

Suezalla fill scale for ggplot
}
