% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{suezalla_fonts}
\alias{suezalla_fonts}
\title{Register fonts for suezalla themes}
\usage{
suezalla_fonts()
}
\description{
Installs and registers required fonts, including Latin Modern Roman, from local files.
}
