% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{suezalla_fonts_zombie}
\alias{suezalla_fonts_zombie}
\title{Register xkcd font for zombie theme}
\usage{
suezalla_fonts_zombie()
}
\description{
You must manually install the xkcd font (e.g., https://github.com/ipython/xkcd-font/raw/master/xkcd-script.ttf)
and ensure it's available in your system fonts folder before calling this.
}
