% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_academic1940}
\alias{theme_academic1940}
\title{Theme: Academic 1940}
\usage{
theme_academic1940(base_family = "abhaya")
}
\arguments{
\item{base_family}{Font to use (default = "abhaya")}
}
\value{
A list: theme and coordinate system
}
\description{
A vintage-style theme inspired by black-and-white academic plots.
}
