% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic_style.R
\name{theme_academic1940_pattern}
\alias{theme_academic1940_pattern}
\title{Academic 1940 Pattern Theme (with default geom_bar_pattern)}
\usage{
theme_academic1940_pattern(
  n = 6,
  pattern_spacing = 0.03,
  pattern_density = NULL,
  pattern_key_scale_factor = 1,
  colour = "black",
  show.legend = TRUE
)
}
\arguments{
\item{n}{Number of pattern levels (i.e., factor levels in fill/pattern)}

\item{pattern_spacing}{Spacing between pattern elements (default 0.03)}

\item{pattern_density}{Pattern density (default NULL = ggpattern default)}

\item{pattern_key_scale_factor}{Size of the pattern in legend (default 1)}

\item{colour}{Outline color (default "black")}

\item{show.legend}{Whether to show legend (default TRUE)}
}
\value{
A list of ggplot2 layers
}
\description{
A vintage-style black-and-white theme for bar plots using ggpattern.
Automatically includes scale_fill_manual, scale_pattern_manual, and a default geom_bar_pattern layer.
}
