% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_magazine}
\alias{theme_magazine}
\title{Magazine Theme}
\usage{
theme_magazine(
  base_size = 12,
  font = c("Roboto", "latinmodern", "abhaya", "CMU Serif")
)
}
\arguments{
\item{base_size}{Base font size}

\item{font}{Character. Font family to use. Options: "Roboto", "latinmodern", "abhaya", "CMU Serif"}
}
\value{
A ggplot2 theme
}
\description{
A clean, minimalistic theme inspired by editorial layouts.
All text in black, no background grids.
}
