% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_suezalla}
\alias{theme_suezalla}
\title{Suezalla Theme}
\usage{
theme_suezalla(base_size = 12, base_family = "Roboto")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family (default = "Roboto")}
}
\value{
A ggplot2 theme
}
\description{
A clean, minimalistic ggplot2 theme. Works best with Roboto (see \code{suezalla_fonts()}).
}
