% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_zombie}
\alias{theme_zombie}
\title{Suezalla Zombie Theme}
\usage{
theme_zombie(base_size = 12, base_family = "xkcd")
}
\value{
A ggplot2 theme object
}
\description{
A playful zombie-style ggplot2 theme for point plots. Requires the XKCD font installed.
}
