% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_xkcd_font}
\alias{install_xkcd_font}
\title{Install xkcd font on macOS, Windows, or Linux}
\usage{
install_xkcd_font()
}
\description{
Downloads and installs the xkcd font to your system font folder.
Required for using \code{theme_zombie()}.
}
