% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_suezalla.R, R/theme_suezalla.R
\name{scale_colour_suezalla}
\alias{scale_colour_suezalla}
\title{Custom ggplot2 colour scale for suezalla palettes}
\usage{
scale_colour_suezalla(...)

scale_colour_suezalla(...)
}
\arguments{
\item{...}{Additional arguments passed to scale functions}

\item{palette_name}{Name of the palette to use}

\item{n}{Number of colours to use}

\item{type}{\"discrete\" or \"continuous\"}
}
\description{
Custom ggplot2 colour scale for suezalla palettes

Suezalla colour scale for ggplot
}
