% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_suezalla.R
\name{scale_fill_suezalla}
\alias{scale_fill_suezalla}
\title{ggplot2 fill scale using suezalla_palette}
\usage{
scale_fill_suezalla(
  palette_name = "polaroid",
  n = NULL,
  type = "discrete",
  ...
)
}
\description{
ggplot2 fill scale using suezalla_palette
}
