% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{suezalla_fonts}
\alias{suezalla_fonts}
\title{Register fonts for suezalla themes}
\usage{
suezalla_fonts()
}
\description{
Loads Roboto and Abhaya Libre from Google Fonts and local LaTeX-like fonts.
You must call this before plotting if you want full font rendering.
}
