% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{suezalla_fonts}
\alias{suezalla_fonts}
\title{Register Roboto font for suezalla theme}
\usage{
suezalla_fonts()
}
\description{
Downloads and registers Roboto from Google Fonts using sysfonts.
You must call this before using \code{theme_suezalla()} if you want custom font rendering.
}
