% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{suezalla_fonts}
\alias{suezalla_fonts}
\title{Suezalla custom ggplot2 theme
Load Roboto font for suezalla theme}
\usage{
suezalla_fonts()
}
\description{
Loads the Roboto Google font and enables \code{showtext} rendering for plots.
Must be called before plotting for the Roboto font to apply.
}
