% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{suezalla_fonts_zombie}
\alias{suezalla_fonts_zombie}
\title{Register xkcd font for suezalla zombie theme}
\usage{
suezalla_fonts_zombie()
}
\description{
Enables rendering with the xkcd font for \code{theme_zombie()}.
Assumes xkcd.ttf is already installed system-wide.
}
