% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{suezalla_fonts_zombie}
\alias{suezalla_fonts_zombie}
\title{Optional helper to manually register the XKCD font}
\usage{
suezalla_fonts_zombie()
}
\description{
Downloads and installs the XKCD font and registers it for use with ggplot2.
}
