% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_academic1940}
\alias{theme_academic1940}
\title{Academic 1940 Theme}
\usage{
theme_academic1940(font = c("abhaya", "CMU Serif", "latinmodern", "Roboto"))
}
\arguments{
\item{font}{Font family (default = "abhaya")}
}
\value{
A list: theme and coord system
}
\description{
A vintage-style theme inspired by mid-century academic plots.
}
