% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_academic1940}
\alias{theme_academic1940}
\title{Theme: Academic 1940}
\usage{
theme_academic1940()
}
\value{
A ggplot2 theme object
}
\description{
A vintage-style ggplot2 theme inspired by classic black-and-white academic plots.
Requires "Abhaya Libre" font (auto-loaded from Google Fonts).
}
