% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic_style.R
\name{theme_academic1940_pattern}
\alias{theme_academic1940_pattern}
\title{Theme: Academic 1940 with Patterns}
\usage{
theme_academic1940_pattern()
}
\value{
A ggplot2 theme object
}
\description{
A vintage-style ggplot2 theme inspired by black-and-white academic plots.
Designed to work with ggpattern fills and the "academic" palette.
Uses the "Abhaya Libre" font if available.
}
