% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_latex}
\alias{theme_latex}
\title{Theme: LaTeX-Style Academic Plot}
\usage{
theme_latex(gridlines = TRUE)
}
\arguments{
\item{gridlines}{Logical. If TRUE (default), shows light dashed major gridlines.}
}
\value{
A ggplot2 theme object
}
\description{
A vintage-inspired ggplot2 theme mimicking LaTeX-style academic figures.
Features serif text (Abhaya Libre), black axis box, and light dashed gridlines (optional).
}
