% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_magazine}
\alias{theme_magazine}
\title{Magazine Theme}
\usage{
theme_magazine(base_size = 12, base_family = "Roboto")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family (default = "Roboto")}
}
\value{
A ggplot2 theme
}
\description{
A clean, minimalistic theme inspired by editorial layouts. All text in black, no background grids.
}
