% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_suezalla.R
\name{theme_suezalla}
\alias{theme_suezalla}
\title{A clean, minimalistic ggplot2 theme designed to work well with Roboto font.
Note: Use \code{suezalla_fonts()} before plotting to load and activate the Roboto font.}
\usage{
theme_suezalla(base_size = 12, base_family = "Roboto")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family, default is "Roboto"}
}
\value{
A ggplot2 theme object
}
\description{
A clean, minimalistic ggplot2 theme designed to work well with Roboto font.
Note: Use \code{suezalla_fonts()} before plotting to load and activate the Roboto font.
}
